DRY fixes
