+++ src/main/webapp/join-room.html
@@ -3,6 +3,7 @@
   <head>
     <meta charset="UTF-8">
     <title>Song Guessing Game</title>
+    <script src="common-functions.js"></script>
     <script src="join-room.js"></script>
   </head>
   <body>
