+++ src/main/webapp/join-room.js
@@ -41,26 +41,9 @@ const joinRoom = async (ev) => {
   window.location.href = url;
 }
 
-function parseRoomId(url) {
-  // Split url at '?' and saves second half (url parameter).
-  let paramString = url.split('?')[1]; 
-  let queryString = new URLSearchParams(paramString);
-
-  // Saves value into roomId (pair[0] would be key).
-  let roomId = "";
-  for (let pair of queryString.entries()) {
-	  roomId = pair[1]; 
-  }
-
-  return roomId;
-}
-
 window.addEventListener('DOMContentLoaded', ()=>{
     // Creates room on button click.
     document.getElementById('create-room-btn').addEventListener('click', createRoom);
     // Joins room on button click.
     document.getElementById('join-room-btn').addEventListener('click', joinRoom);
 });
-
-// Testing exports
-exports.parseRoomId = parseRoomId;
