+++ src/main/webapp/lobby.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <title>Song Guessing Game</title>
     <script src="https://js.pusher.com/6.0/pusher.min.js"></script>
+    <script src="common-functions.js"></script>
     <script src="lobby.js"></script>
   </head>
   <body onload="loadUsernames()">
