+++ src/main/webapp/lobby.js
@@ -44,20 +44,6 @@ function createUsernameElement(username) {
   return node;
 }
 
-function parseRoomId(url) {
-  // Split url at '?' and saves second half (url parameter).
-  let paramString = url.split('?')[1]; 
-  let queryString = new URLSearchParams(paramString);
-
-  // Saves value into roomId (pair[0] would be key).
-  let roomId = "";
-  for (let pair of queryString.entries()) {
-	  roomId = pair[1]; 
-  }
-
-  return roomId;
-}
-
 window.addEventListener('DOMContentLoaded', ()=>{
     // Displays url on lobby page.
     document.getElementById("url").innerHTML = "Share the link!:<br>" + (window.location.href);
